Added Panoptic Colmap helper

+++ script/pre_panoptic.py
@@ -1,32 +1,7 @@
-# MIT License
-
-# Copyright (c) 2023 OPPO
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
 import os 
-import cv2 
 import glob 
-import tqdm 
-import numpy as np 
 import shutil
-import pickle
+import numpy as np 
 import sys 
 import argparse
 sys.path.append(".")
@@ -34,39 +9,9 @@ from thirdparty.gaussian_splatting.utils.my_utils import posetow2c_matrcs, rotma
 from thirdparty.colmap.pre_colmap import * 
 from thirdparty.gaussian_splatting.helper3dg import getcolmapsinglen3d
 
-
-
-
-# def extractframes(videopath):
-#     cam = cv2.VideoCapture(videopath)
-#     ctr = 0
-#     sucess = True
-#     for i in range(300):
-#         if os.path.exists(os.path.join(videopath.replace(".mp4", ""), str(i) + ".png")):
-#             ctr += 1
-#     if ctr == 300 or ctr == 150: # 150 for 04_truck 
-#         print("already extracted all the frames, skip extracting")
-#         return
-#     ctr = 0
-#     while ctr < 300:
-#         try:
-#             _, frame = cam.read()
-
-#             savepath = os.path.join(videopath.replace(".mp4", ""), str(ctr) + ".png")
-#             if not os.path.exists(videopath.replace(".mp4", "")) :
-#                 os.makedirs(videopath.replace(".mp4", ""))
-
-#             cv2.imwrite(savepath, frame)
-#             ctr += 1 
-#         except:
-#             sucess = False
-#             print("error")
-#     cam.release()
-#     return
-
-
-def preparecolmapdynerf(folder, offset=0):
-    folderlist = glob.glob(folder + "cam**/")
+def preparecolmappanoptic(folder, offset=0):
+    # Only process folders inside `ims` and exclude `seg`
+    folderlist = sorted(glob.glob(os.path.join(folder, "ims", "*/")))
     imagelist = []
     savedir = os.path.join(folder, "colmap_" + str(offset))
     if not os.path.exists(savedir):
@@ -74,23 +19,22 @@ def preparecolmapdynerf(folder, offset=0):
     savedir = os.path.join(savedir, "input")
     if not os.path.exists(savedir):
         os.mkdir(savedir)
-    for folder in folderlist :
-        imagepath = os.path.join(folder, str(offset) + ".png")
-        imagesavepath = os.path.join(savedir, folder.split("/")[-2] + ".png")
 
-        shutil.copy(imagepath, imagesavepath)
+    for camera_folder in folderlist:
+        imagepath = os.path.join(camera_folder, f"{str(offset).zfill(6)}.jpg")  # Handle `.jpg`
+        imagesavepath = os.path.join(savedir, camera_folder.split("/")[-2] + ".jpg")  # Save as `.jpg`
 
+        if os.path.exists(imagepath):
+            shutil.copy(imagepath, imagesavepath)
+        else:
+            print(f"Image {imagepath} not found. Skipping.")
 
-    
-def convertdynerftocolmapdb(path, offset=0):
+def convertpanoptictocolmapdb(path, offset=0):
     originnumpy = os.path.join(path, "poses_bounds.npy")
-    video_paths = sorted(glob.glob(os.path.join(path, 'cam*.mp4')))
+    camera_folders = sorted(glob.glob(os.path.join(path, 'ims', '*/')))  # Camera folders inside 'ims'
     projectfolder = os.path.join(path, "colmap_" + str(offset))
-    #sparsefolder = os.path.join(projectfolder, "sparse/0")
     manualfolder = os.path.join(projectfolder, "manual")
 
-    # if not os.path.exists(sparsefolder):
-    #     os.makedirs(sparsefolder)
     if not os.path.exists(manualfolder):
         os.makedirs(manualfolder)
 
@@ -99,25 +43,23 @@ def convertdynerftocolmapdb(path, offset=0):
     savepoints = os.path.join(manualfolder, "points3D.txt")
     imagetxtlist = []
     cameratxtlist = []
+
     if os.path.exists(os.path.join(projectfolder, "input.db")):
         os.remove(os.path.join(projectfolder, "input.db"))
 
     db = COLMAPDatabase.connect(os.path.join(projectfolder, "input.db"))
-
     db.create_tables()
-
 
     with open(originnumpy, 'rb') as numpy_file:
         poses_bounds = np.load(numpy_file)
         poses = poses_bounds[:, :15].reshape(-1, 3, 5)
 
-        llffposes = poses.copy().transpose(1,2,0)
+        llffposes = poses.copy().transpose(1, 2, 0)
         w2c_matriclist = posetow2c_matrcs(llffposes)
-        assert (type(w2c_matriclist) == list)
-
+        assert isinstance(w2c_matriclist, list)
 
         for i in range(len(poses)):
-            cameraname = os.path.basename(video_paths[i])[:-4]#"cam" + str(i).zfill(2)
+            cameraname = os.path.basename(camera_folders[i])[:-1]  # Remove trailing slash
             m = w2c_matriclist[i]
             colmapR = m[:3, :3]
             T = m[:3, 3]
@@ -125,101 +67,71 @@ def convertdynerftocolmapdb(path, offset=0):
             H, W, focal = poses[i, :, -1]
             
             colmapQ = rotmat2qvec(colmapR)
-            # colmapRcheck = qvec2rotmat(colmapQ)
 
-            imageid = str(i+1)
+            imageid = str(i + 1)
             cameraid = imageid
-            pngname = cameraname + ".png"
+            jpgname = cameraname + ".jpg"  # Handle `.jpg`
             
-            line =  imageid + " "
-
-            for j in range(4):
-                line += str(colmapQ[j]) + " "
-            for j in range(3):
-                line += str(T[j]) + " "
-            line = line  + cameraid + " " + pngname + "\n"
-            empltyline = "\n"
+            line = imageid + " "
+            line += " ".join(map(str, colmapQ)) + " "
+            line += " ".join(map(str, T)) + " "
+            line += cameraid + " " + jpgname + "\n"
             imagetxtlist.append(line)
-            imagetxtlist.append(empltyline)
+            imagetxtlist.append("\n")
 
             focolength = focal
-            model, width, height, params = i, W, H, np.array((focolength,  focolength, W//2, H//2,))
+            model, width, height, params = i, W, H, np.array((focolength, focolength, W // 2, H // 2))
 
             camera_id = db.add_camera(1, width, height, params)
-            cameraline = str(i+1) + " " + "PINHOLE " + str(width) +  " " + str(height) + " " + str(focolength) + " " + str(focolength) + " " + str(W//2) + " " + str(H//2) + "\n"
+            cameraline = f"{i + 1} PINHOLE {width} {height} {focolength} {focolength} {W // 2} {H // 2}\n"
             cameratxtlist.append(cameraline)
             
-            image_id = db.add_image(pngname, camera_id,  prior_q=np.array((colmapQ[0], colmapQ[1], colmapQ[2], colmapQ[3])), prior_t=np.array((T[0], T[1], T[2])), image_id=i+1)
+            db.add_image(jpgname, camera_id, prior_q=colmapQ, prior_t=T, image_id=i + 1)
             db.commit()
-        db.close()
 
+    db.close()
 
     with open(savetxt, "w") as f:
-        for line in imagetxtlist :
-            f.write(line)
+        f.writelines(imagetxtlist)
     with open(savecamera, "w") as f:
-        for line in cameratxtlist :
-            f.write(line)
+        f.writelines(cameratxtlist)
     with open(savepoints, "w") as f:
-        pass 
+        pass
 
-
-
-
-
-if __name__ == "__main__" :
+if __name__ == "__main__":
     parser = argparse.ArgumentParser()
- 
-    parser.add_argument("--videopath", default="", type=str)
+    parser.add_argument("--folderpath", default="", type=str)
     parser.add_argument("--startframe", default=0, type=int)
     parser.add_argument("--endframe", default=150, type=int)
 
     args = parser.parse_args()
-    videopath = args.videopath
-
+    folderpath = args.folderpath
     startframe = args.startframe
     endframe = args.endframe
 
-
     if startframe >= endframe:
-        print("start frame must smaller than end frame")
+        print("start frame must be smaller than end frame")
         quit()
     if startframe < 0 or endframe > 300:
-        print("frame must in range 0-300")
+        print("frame must be in range 0-300")
         quit()
-    if not os.path.exists(videopath):
-        print("path not exist")
+    if not os.path.exists(folderpath):
+        print("path does not exist")
         quit()
     
-    if not videopath.endswith("/"):
-        videopath = videopath + "/"
-    
-    
-    
-    ##### step1
-    print("start extracting 300 frames from videos")
-    videoslist = glob.glob(videopath + "*.mp4")
-    for v in tqdm.tqdm(videoslist):
-        extractframes(v)
+    if not folderpath.endswith("/"):
+        folderpath = folderpath + "/"
 
-    
-
-    # # ## step2 prepare colmap input 
-    print("start preparing colmap image input")
+    # Step 1: Prepare colmap input (skip frame extraction)
+    print("Start preparing colmap image input")
     for offset in range(startframe, endframe):
-        preparecolmapdynerf(videopath, offset)
+        preparecolmappanoptic(os.path.join(folderpath, "ims"), offset)
 
-
-    print("start preparing colmap database input")
-    # # ## step 3 prepare colmap db file 
+    # Step 2: Prepare colmap database input
+    print("Start preparing colmap database input")
     for offset in range(startframe, endframe):
-        convertdynerftocolmapdb(videopath, offset)
+        convertpanoptictocolmapdb(os.path.join(folderpath, "ims"), offset)
 
-
-    # ## step 4 run colmap, per frame, if error, reinstall opencv-headless 
+    # Step 3: Run colmap, per frame
     for offset in range(startframe, endframe):
-        getcolmapsinglen3d(videopath, offset)
-
-
-
-
+        getcolmapsinglen3d(folderpath, offset)
